Use correct flags in component customizer
In kubebuilder the manager flags `--metrics-addr` and `enable-leader-election`
were deprecated in favor of --metrics-bind-address and --leader-elect to be more
aligned with core Kubernetes Components.

For more info:
https://github.com/kubernetes-sigs/kubebuilder/issues/1839

All cluster api providers have made related changes. It leads to the fact
that when we try to use old flags there, operator returns an error.