fix: Avoid failure on unchanged provider version
In the current implementation, when a provider version remains the same
during subsequent reconciliations, the operator encounters an error with
the message "failed getting clusterctl Provider version," leading to a
degraded status.

This PR addresses this issue by adding a check to determine whether
the provider version has changed or not.