fix: prevent implicit memory aliasing in for loops
In golang the iteration variable is reused. At each iteration, the value
of the next element in the range expression is assigned to the iteration
variable; "provider" doesn't change, only its value changes. Hence, the
expression &provider is referring to the same location in memory.

The recommended solution for the issue is to redeclare "provider" inside
the for loop.

For more information:
https://github.com/golang/go/discussions/56010
https://husni.dev/beware-of-implicit-memory-aliasing-in-go-foor-loop/