chore: allow to halt reconciliation
Currently in case of successful execution of a phase, operator will always
start execution of the next one, and the only way to stop reconciliation is
to return an error from a phase function. But sometimes we need just to
stop it without any errors. For example, when provider version didn't change.

This commit introduces a halt flag for the phase functions as a return value.
If the flag set to true by a function, operator stops reconciling the provider
immediately.