docs: Remove outdated CAPI operator proposal
The original CAPI operator proposal has become outdated and is no
longer relevant as documentation. It was intended to outline project
goals, not describe current features. As new documentation has been
added, there is considerable overlap, causing potential confusion
for users. To prevent this, we are deleting the outdated proposal.

A copy of the proposal is available in the CAPI repo:
https://github.com/kubernetes-sigs/cluster-api/blob/main/docs/proposals/20201020-capi-provider-operator.md