fix: delete all provider resources
Client proxy defines among which types we will search for objects with
the desired labels. Currently this list is not full and we skip some
types, like ClusterRole or ClusterRole bindings. It leads to the fact
that objects of these types stay in the system when their provider is
deleted.

This PR updates the list of resource types we need to scan.