fix: compress components manifests if they don't fit the configmap
By default we download capi provider manifests from a provided URL
and store it in a confimap to reuse later. Currently there is a limit
on the configmap size, which is 1mb. If we the manifests size is bigger,
then the installation process fails.

To avoid it we will compress manifests if they can't be stored in the
configmap directly.