chore: add a preflight check to validate core provider name
When you try to install a Core Provider with name different from
"cluster-api", clusterctl validation fails:
https://github.com/kubernetes-sigs/cluster-api/blob/main/cmd/clusterctl/client/config/providers_client.go#L414-L416
After that CAPI operator puts a vague error description in the logs
and doesn't add anything meaningful to the provider status.

To improve it, we add a new pre-flight check to compare Core Provider
name with the desired one, and fail earlier, updating the status with
correct information.