Get the latest operator version from the list of all available versions
Now we use "latest" tag, but due to recent clusterctl changes it forces
to download metadata.yaml file which is not available for CAPI operator.
Because of that installation fails with exceeded context deadline.

To prevent this we specify a version when we create a repo object, and
then pick the latest version from the list of all available versions.