Add internal controllers to public controller/alias.go
This change adds existing internal controller implementations into the `controller/alias.go` file.

This is an initial step towards addressing issue #798, which aims to provide a public API for CAPI Operator providers. The current controller implementation is internal, and this change is a step towards organizing the internal structure, potentially in preparation for exposing parts of it or providing extension points for custom providers.

Signed-off-by: Danil-Grigorev <danil.grigorev@suse.com>